Abstracting getIsDocsLink helper
