+++ src/lib/remark-plugins/rewrite-tutorial-links/utils/index.ts
@@ -1,3 +1,4 @@
+export { getIsDocsLink } from './get-is-docs-link'
 export { getIsLearnLink } from './get-is-learn-link'
 export { getTutorialMap } from './get-tutorial-map'
 export { handleCollectionLink } from './handle-collection-link'
